% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuStep.R
\name{simuStep}
\alias{simuStep}
\title{simuStep function}
\usage{
simuStep(n)
}
\description{
This function allows you to simulate a random process of walking.
Assuming you have equally probability to take your next step towards
all the directions, and then another step again. The output will be a
figure show all the foot prints.
}
\examples{
simuStep(10000) # assuming 10000 steps

}
\keyword{simuStep}
